Test List Endpoint Query in Mosaic Production
This commit adds changes found while testing the Mosaic production API:
- Creates a dbo schema for the test database to match structure of production DB.
- Changes data types & lengths in infrastucture columns to match the ones expected in the production DB.
- Refactors the get all residents gateway method so that querying is quicker and returns distinct data.
- Adds missing Mosaic Id in response object.

To Do: In a separate PR, will add missing phone types that are present in the DB but have not be added as enum values in the code base.

+++ MosaicResidentInformationApi/V1/Gateways/IMosaicGateway.cs
@@ -7,6 +7,6 @@ namespace MosaicResidentInformationApi.V1.Gateways
     {
         List<ResidentInformation> GetAllResidents(int cursor, int limit, string firstname = null, string lastname = null,
             string postcode = null, string address = null);
-        ResidentInformation GetEntityById(int id);
+        ResidentInformation GetEntityById(long id);
     }
 }
